Move digraph_union to union
