+++ src/lib.rs
@@ -47,6 +47,7 @@ use simple_path::*;
 use transitivity::*;
 use traversal::*;
 use tree::*;
+use union::*;
 
 use hashbrown::{HashMap, HashSet};
 
@@ -71,49 +72,6 @@ use crate::generators::PyInit_generators;
 
 pub trait NodesRemoved {
     fn nodes_removed(&self) -> bool;
-}
-
-/// Return a new PyDiGraph by forming a union from two input PyDiGraph objects
-///
-/// The algorithm in this function operates in three phases:
-///
-///  1. Add all the nodes from  ``second`` into ``first``. operates in O(n),
-///     with n being number of nodes in `b`.
-///  2. Merge nodes from ``second`` over ``first`` given that:
-///
-///     - The ``merge_nodes`` is ``True``. operates in O(n^2), with n being the
-///       number of nodes in ``second``.
-///     - The respective node in ``second`` and ``first`` share the same
-///       weight/data payload.
-///
-///  3. Adds all the edges from ``second`` to ``first``. If the ``merge_edges``
-///     parameter is ``True`` and the respective edge in ``second`` and
-///     first`` share the same weight/data payload they will be merged
-///     together.
-///
-///  :param PyDiGraph first: The first directed graph object
-///  :param PyDiGraph second: The second directed graph object
-///  :param bool merge_nodes: If set to ``True`` nodes will be merged between
-///     ``second`` and ``first`` if the weights are equal.
-///  :param bool merge_edges: If set to ``True`` edges will be merged between
-///     ``second`` and ``first`` if the weights are equal.
-///
-///  :returns: A new PyDiGraph object that is the union of ``second`` and
-///     ``first``. It's worth noting the weight/data payload objects are
-///     passed by reference from ``first`` and ``second`` to this new object.
-///  :rtype: PyDiGraph
-#[pyfunction]
-#[pyo3(text_signature = "(first, second, merge_nodes, merge_edges, /)")]
-fn digraph_union(
-    py: Python,
-    first: &digraph::PyDiGraph,
-    second: &digraph::PyDiGraph,
-    merge_nodes: bool,
-    merge_edges: bool,
-) -> PyResult<digraph::PyDiGraph> {
-    let res =
-        union::digraph_union(py, first, second, merge_nodes, merge_edges)?;
-    Ok(res)
 }
 
 /// Color a PyGraph using a largest_first strategy greedy graph coloring.
